move package object to utils folder
